code/requirements.txt
numpy==1.18.1
opencv==3.4.2
python==3.6.0
tensorflow-gpu==1.9.0
